split out UnusedTrainCarDeleter patch + fix string.Format bug
